requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-huggingface
langchain-ollama
langchain-qdrant
sentence-transformers
azure-storage-blob
qdrant-client
pypdf
python-dotenv